add basic support for extra channels
This is just a basic implementation for the primitive image object and
has little support in file handlers or higher level functions (like
cloning images.)

Later changes (much later) will add support to file reading to support
extra channels, and even later will add support for reading non-RGBA
colorspaces into extra-channel-only images.

The main aim here is to get extra channels into the ABI to get a few
ABI-breaking changes out of the way in a single release.

Code that directly accesses image data will need to be updated, but
there's other changes in this batch that should deal with that.

+++ tags.c
@@ -668,7 +668,7 @@ void i_tags_print(i_img_tags *tags) {
     printf(" Code : %d\n", tag->code);
     if (tag->data) {
       int pos;
-      printf(" Data : %d (%p) => '", tag->size, tag->data);
+      printf(" Data : %lu (%p) => '", (unsigned long)tag->size, tag->data);
       for (pos = 0; pos < tag->size; ++pos) {
 	if (tag->data[pos] == '\\' || tag->data[pos] == '\'') {
 	  putchar('\\');
